chore: Update Ruby version to 3.0 and install Fastlane
Update the GitHub Actions workflow to include the setup of Ruby version 3.0 and the installation of Fastlane. This will ensure that the project is using the latest stable version of Ruby and has Fastlane installed for automated iOS and Android builds.

Bump iOS build number to 1 and Android version code to 1

Update the app.json file to bump the iOS build number to 1 and the Android version code to 1. This is necessary to ensure that the app can be properly identified and updated on both iOS and Android devices.

Update GitHub Actions workflow to use Ubuntu latest runner for Android build

Update the GitHub Actions workflow to use the latest version of the Ubuntu runner for Android builds. This will ensure that the build environment is up to date and compatible with the latest Android SDK and tools.

Bump iOS build number to 11 and Android version code to 11

Bump iOS build number to 10 and Android version code to 10

Bump iOS build number to 9 and Android version code to 9

Bump iOS build number to 8 and Android version code to 8